refactor: 클래스명 + Test 형태로 test class 분리
