+++ src/test/java/racingcar/RandomNumberGeneratorTest.java
@@ -1,0 +1,17 @@
+package racingcar;
+
+import org.junit.jupiter.api.Test;
+import racingcar.util.RandomNumberGenerator;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+public class RandomNumberGeneratorTest {
+    private static final int TESTCASE_NUM = 10000;
+
+    @Test
+    void 랜덤_숫자_생셩() {
+        for (int i = 0; i < TESTCASE_NUM; i++) {
+            assertThat(RandomNumberGenerator.generate()).isBetween(0,9);
+        }
+    }
+}
